+++ vscode-extension/src/test/runTests.ts
@@ -1,5 +1,6 @@
 import { runComposerPromptTests } from "./composerPrompt.test";
 import { runDiagnosticRulesTests } from "./diagnosticRules.test";
+import { runFeedbackLinksTests } from "./feedbackLinks.test";
 import { runGeneratedDocumentsTests } from "./generatedDocuments.test";
 import { runPolicyDocumentTests } from "./policyDocument.test";
 import { runRepoResolverTests } from "./repoResolver.test";
@@ -7,6 +8,7 @@ import { runReleaseReadinessTests } from "./releaseReadiness.test";
 import { runSetupDoctorTests } from "./setupDoctor.test";
 import { runTelemetryStoreTests } from "./telemetryStore.test";
 import { runTemplateDocumentTests } from "./templateDocument.test";
+import { runWelcomeContentTests } from "./welcomeContent.test";
 import { runWorkspaceLintRulesTests } from "./workspaceLintRules.test";
 
 /**
@@ -15,6 +17,7 @@ import { runWorkspaceLintRulesTests } from "./workspaceLintRules.test";
 function main(): void {
   runComposerPromptTests();
   runDiagnosticRulesTests();
+  runFeedbackLinksTests();
   runGeneratedDocumentsTests();
   runTemplateDocumentTests();
   runWorkspaceLintRulesTests();
@@ -23,6 +26,7 @@ function main(): void {
   runReleaseReadinessTests();
   runSetupDoctorTests();
   runTelemetryStoreTests();
+  runWelcomeContentTests();
 
   if (process.exitCode) {
     return;